test-seccomp: Fix function argument order
The seccomp filter looks at the bits in the mode_t (the third argument,
SCMP_A2(*)), which is correct. The test, however, got 2 arguments
swapped in the ad-hoc syscall prototype. They are still called in the
correct order, though.

The ad-hoc prototype was destined to be replaced by a normal
missing-syscall prototype anyway, but we still fix this in a separate
patch, so in the likely case PR 29618 hits systemd-stable this can be
applied on top without pulling future unrelated changes.

Fixes: 8b45281daa3a ("seccomp: also check the mode parameter of `fchmodat2(2)`")
Link: https://github.com/torvalds/linux/blob/v6.6-rc6/fs/open.c#L701